Address review feedback on protocol types

+++ Minecraft.World/Network/Packets/KickPlayerPacket.cpp
@@ -23,7 +23,7 @@ void KickPlayerPacket::handle(PacketListener *listener)
 
 void KickPlayerPacket::read(DataInputStream *dis) //throws IOException 
 {
-	m_networkSmallId = static_cast<std::uint8_t>(dis->readByte());
+	m_networkSmallId = dis->readByte();
 }
 
 void KickPlayerPacket::write(DataOutputStream *dos) //throws IOException
